robot_arm/requirements.txt
opencv-python
numpy
pyserial
scipy
